# [FIX] stock: improve move_line._free_reservation performances