Move candidate.unlink out of the for loop to unlink records in batch.

Turn self.mapped.filtered -> self.search to execute both in one query to the DB.

##### speed-up

Customer DB, inventory with 532 lines

Total inventory validation time: `18min -> 10s`

`move_line._free_reservation` times when validating inventory

| | Before PR | After PR|
|:--:|:------------:|:-----------:|
| avg | 9.48s | 0.067s |
| min | 0.001s | 0.001s |
| max | 16min | 5.38s |

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr