@Aurelienvd 2021-07-02
@ryv-odoo Sure.